Switch to CMake and KDevelop4

+++ include/astro/util.h
@@ -46,9 +46,9 @@ namespace util {
 	inline std::string toupper(const std::string &s) { std::string o(s); FOREACH2(std::string::iterator, o) { *i = ::toupper(*i); }; return o; }
 
 	/// convert long long to std::string
-	inline std::string str(long long n) { char buf[20]; sprintf(buf, "%ld", n); return buf; }
-	/// convert size_t to std::string
-	inline std::string str(size_t n) { char buf[20]; sprintf(buf, "%d", n); return buf; }
+	inline std::string str(long long n) { char buf[20]; sprintf(buf, "%lld", n); return buf; }
+	/// convert unsigned long long to std::string
+	inline std::string str(unsigned long long n) { char buf[20]; sprintf(buf, "%llu", n); return buf; }
 	/// convert int to std::string
 	inline std::string str(int n) { char buf[20]; sprintf(buf, "%d", n); return buf; }
 	/// convert unsigned int to std::string
